pyproject.toml
[project]
name = "fastapi-template"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi==0.115.12",
    "imported==0.1.1",
    "pydantic-settings>=2.9.1",
    "uvicorn==0.34.2",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "mypy==1.15.0",
    "pytest==8.3.5",
    "pytest-dotenv>=0.5.2",
    "python-dotenv>=1.1.0",
    "ruff==0.11.7",
]

[tool.commitizen]
name = "cz_conventional_commits"
tag_format = "$version"
version_scheme = "semver2"
version_provider = "uv"
update_changelog_on_bump = true
major_version_zero = true
version_files = [
    "app/__version__.py"
]

[tool.pytest.ini_options]
env_files = [
    ".env.development",
]